excessões 1

+++ src/alunos/alunos.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateAlunoDto } from './dto/create-aluno.dto';
 import { UpdateAlunoDto } from './dto/update-aluno.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Aluno } from './entities/aluno.entity';
 import { Repository } from 'typeorm';
+import { AlunoNotFoundException } from './exceptions/alunoNotFound.exception';
 
 @Injectable()
 export class AlunosService {
@@ -25,7 +26,17 @@ export class AlunosService {
   }
 
   async findOne(id: number) {
-    return await this.repository.findOneBy({ id: id });
+    try {
+      return await this.repository.findOneByOrFail({ id: id });
+    } catch (error) {
+      if (error.name === 'EntityNotFoundError') {
+        throw new AlunoNotFoundException();
+      }
+      throw new HttpException(
+        'Erro interno de servidor',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
   }
 
   async update(id: number, updateAlunoDto: UpdateAlunoDto) {
@@ -33,6 +44,10 @@ export class AlunosService {
   }
 
   async remove(id: number) {
+    const response = await this.repository.delete({ id: id });
+    if (response.affected === 0) {
+      throw new AlunoNotFoundException();
+    }
     return await this.repository.delete({ id: id });
   }
 }
